fix(dictionary): the word page marks a root it only guessed at
The popup has marked rule-derived roots since they were introduced (27ef7f9),
and the word page never did: it renders the chain's indent and drops the
warning, so a demutation the lemma table worked out by rule reads exactly
like something Cregeen wrote down. 'aarkys' offers 'cha n'aark' as its root
with nothing to say the link is a guess.

The mark moves out of the popup to reach the page, and the entry helpers go
with it: the page already imported headingFor from a modal component, which
is not where a rule about dictionary entries lives.

The wording loses "the word you tapped": the page is not tapped.
