+++ CorpusSearch/ClientApp/src/components/DictionaryLookupModal.css
@@ -49,22 +49,6 @@
     font-size: 0.9em;
 }
 
-/* a root reached only by rule: the dictionary documents the entry, but nothing
-   documents it as this word's root, so it may simply be wrong. Wears the
-   suggestion box's amber, the popup's existing colour for "this is a guess" */
-.dict-popup-unverified {
-    background: #fff8e1;
-    border: 1px solid #ecdfae;
-    color: #6b5a1e;
-    font-size: 0.8em;
-    font-weight: 600;
-    border-radius: 4px;
-    padding: 0 5px;
-    margin-left: 4px;
-    text-decoration: none;
-    white-space: nowrap;
-}
-
 /* the "did you mean" fallback: a tinted suggestion box, visually apart from
    real entries - these words are guesses, not what the text says */
 .dict-popup-suggestions {
